use BLL to request

+++ WebService/BLL/IPersonManager.cs
@@ -11,7 +11,7 @@ namespace BLL
     {
         Person AddPerson(Person newPerson);
 
-        Person UpdatePersonQuota(int id, float quota);
+        Person UpdatePersonQuota(int id, double quota);
 
         Person getPersonByUID(String UID);
